feat: change loading on splash screen to MainActivity
