+++ app/src/main/java/il/androidcourse/infonow/RSSWorker.java
@@ -107,7 +107,7 @@ public class RSSWorker extends Worker {
     private Notification createNotification(String title, Bitmap image) {
 
         Context context = getApplicationContext();
-        Intent intent = new Intent(context, SplashActivity.class);
+        Intent intent = new Intent(context, MainActivity.class);
         intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
         PendingIntent pendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_IMMUTABLE);
