+++ app/src/main/res/layout/activity_main.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -8,36 +8,70 @@
 
     <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/menu_bar">
+        android:layout_height="match_parent"
+        app:layout_constraintTop_toTopOf="parent">
 
-        <ImageButton
-            android:id="@+id/home_button"
-            android:layout_width="64dp"
-            android:layout_height="64dp"
-            android:scaleType="centerCrop"
-            android:layout_centerInParent="true"
-            android:src="@drawable/i"
-            android:background="@null"/>
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/menu_bar">
 
-        <ImageButton
-            android:id="@+id/settings_button"
-            android:src="@drawable/settings"
-            android:layout_width="64dp"
-            android:layout_height="64dp"
-            android:layout_centerVertical="true"
-            app:tint="@color/black"
-            android:scaleType="centerCrop"
-            android:layout_alignParentRight="true"
-            android:background="@null"/>
-<!--        make tint for icons depend on -->
+            <ImageButton
+                android:id="@+id/home_button"
+                android:layout_width="64dp"
+                android:layout_height="64dp"
+                android:layout_centerInParent="true"
+                android:background="@null"
+                android:scaleType="centerCrop"
+                android:src="@drawable/i" />
 
+            <ImageButton
+                android:id="@+id/settings_button"
+                android:layout_width="64dp"
+                android:layout_height="64dp"
+                android:layout_alignParentRight="true"
+                android:layout_centerVertical="true"
+                android:background="@null"
+                android:scaleType="centerCrop"
+                android:src="@drawable/settings"
+                app:tint="@color/black" />
+            <!--        make tint for icons depend on -->
+
+        </RelativeLayout>
+
+        <!-- Fragment Container -->
+        <FrameLayout
+            android:id="@+id/fragment_container"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_below="@id/menu_bar"/>
     </RelativeLayout>
 
-    <!-- Fragment Container -->
-    <FrameLayout
-        android:id="@+id/fragment_container"
+    <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_below="@id/menu_bar"/>
-</RelativeLayout>
+        android:id="@+id/splash"
+        android:background="#FFFFFF">
+
+        <ImageView
+            android:layout_centerInParent="true"
+            android:id="@+id/imageView"
+            android:layout_width="164dp"
+            android:layout_height="164dp"
+            android:src="@drawable/i"
+            android:scaleType="centerCrop"
+            android:background="@null"/>
+
+        <ProgressBar
+            android:id="@+id/progressBar"
+            android:layout_width="128dp"
+            android:layout_height="16dp"
+            android:layout_below="@id/imageView"
+            android:layout_marginTop="16dp"
+            android:layout_centerInParent="true"
+            style="@style/Widget.AppCompat.ProgressBar.Horizontal"
+            android:indeterminate="false"
+            android:progress="0"
+            android:progressBackgroundTint="#597CFF"/>
+    </RelativeLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
