--- app/src/main/res/layout/activity_splash.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".SplashActivity">
-
-    <ImageView
-        android:layout_centerInParent="true"
-        android:id="@+id/imageView"
-        android:layout_width="164dp"
-        android:layout_height="164dp"
-        android:src="@drawable/i"
-        android:scaleType="centerCrop"
-        android:background="@null"/>
-
-    <ProgressBar
-        android:id="@+id/progressBar"
-        android:layout_width="128dp"
-        android:layout_height="16dp"
-        android:layout_below="@id/imageView"
-        android:layout_marginTop="16dp"
-        android:layout_centerInParent="true"
-        style="@style/Widget.AppCompat.ProgressBar.Horizontal"
-        android:indeterminate="false"
-        android:progress="0"
-        android:progressBackgroundTint="#597CFF"/>
-</RelativeLayout>
